Conflicts:
	src/engine/lib/GUIDOPianoRoll.cpp
	src/engine/pianoroll/PianoRoll.cpp
	src/engine/pianoroll/PianoRollTrajectory.h
	src/engine/pianoroll/ReducedProportional.cpp

+++ src/engine/abstract/Fraction.h
@@ -72,8 +72,8 @@ class Fraction
 		Fraction & operator *=(const int tmp);
 
 	private:
-		int numerator;
-		int denominator;
+		long long numerator;
+		long long denominator;
 		double dval;
 };
